refactor: 함수명, 변수명 일부 수정
