+++ src/index.ts
@@ -1,9 +1,9 @@
 import { renderInputSection } from './view/renderer.js';
-import { setInputButtonsEventListener } from './controller/event.js';
+import { setInputButtonsEvent } from './controller/event.js';
 
 const app = (): void => {
   renderInputSection();
-  setInputButtonsEventListener();
+  setInputButtonsEvent();
 };
 
 app();
